Fixes ACRO bug, starts moving to class based
The acro bug was in the definintions of acronyms in the
\DeclareAcronym command; I had the command syntax wrong.
With the incorrect syntax, the TeX document would not compile.
So, the regex pattern has been updated to parse the correct syntax.

Am moving to a LaTeX class, where a file path is taken as input
and methods will be used to compile, run latexdiff, and convert to
docx. This should clean up code nicely.

v0.3.0