Fixes regex matching in acronyms
With previous 'fix' for acronyms definition parsing, faild to
update the referenced capture group to be group 2 on the recursive
search; it must be 2 because the first capture group is now the
acronym key definition. This has been done through the addition of
a group keyword to the recursiveRegex function. The default value
for group is 1, but for acronyms is set to 2.

v0.3.2