Adds autoCompile CLI; not work with vim
The autoCompile CLI has been added, which is a watchdog that
watches a user entered directory for modifications to .tex files.
When modified, the document will be recompiled. Note that this
does not currently work for files in vim as vim uses .swp files
to track changes in open documents. Thus, a vim parser (if exists)
might be used to get this to work.

Also removes error line in  trackChanges method of LaTeX class
as not really needed.

v0.4.0