Improve compile logic for fast compile
When compiling the document, now checks for existing .aux file.
If found, file is read in.
After first run of TeX compile, the old aux file data, if it
exists, is checked agains the new aux file data. If the files
match, no more runs of the compiler are required as no references
have changed.

v0.4.1