Fixes infinite recursion issue
When adding trackChanges() and toDOCX() method class to the
compile() method, this created infinite recursion with
trackChanges(), as that method calls compile().

This has been fixed by removing those lines and adding them back
into the compileLaTeX CLI; I knew there was a reason I had it this
way, but I was trying to be smart.

v0.4.3