Updates autoCompile to be more efficient
The autoCompile watchdog on linux was running very often as the
swap file for vim changed a lot triggering almost constant
recompiles. This may be an issue on other operating systems too,
but they have not been tested.

An event is also added to signal that the file is being compiled.
This is done so that the compile code will not happen during
a running compile.

v0.5.3