Enables !TEX root definition, chapter references
The compiler will now search for the !TEX root definition in files
in attempt to find root file. If writing a multifile document and
try to compile one of the sub-tex files, an error will occur. This
could be a big issue with the autoCompile CLI. So, now try to find
root TeX file, and compile that instead of document changed, or
document requested by user.

Adds support for references per chapter compiling. Old versions
simply ran BibTeX on the .aux file with name matching the main
tex document. Now, will, recursivley, search directory for all
.aux files, running BibTeX on each file.

v0.6.0