Fixes bugs in include/input and track changes
There was an issue with the regex for finding \include \input
commands which seemed to lead to child files not being included.
This has been fixed.

This fix also includes importing all the child files by default
so that the loaded text is one complete document.

Track changes if updated so that docements with child files actually
produce correct tracked changes. This was done by creating a new
gitShow() function that will get text of given file from given
branch. This new function is now used when initializing the
LaTeXBase class for file loading. If the new gitBranch keyword is
passed on class initialization, then the requested file from the
given branch, all with any children from that branch, are loaded.
When running latexdiff, a temporary  'reference' file, with all child
files loaded into the single file, is created. A new file is created
with the current data (all child files loaded) as well. These two
files are then loaded in latex diff.

v0.7.0