Fixes texlive version not passed to latexdiff
There was a bug where the texlive version specified by user was
not passed to the latexdiff commands, so compilation would fail.
New version fixes this issue by setting the texlive version at
the class level (done in LaTeXBase). The XELATEX, PDFLATEX, and
BIBTEX class attributes in the LaTeX class are now properties that
are updated on call to use the current texlive version set by the
self.texlive attribute.

Updates comment blocks to be more informative or exist.

v0.9.0