Greatly improve acronym substitutions
There were some bugs with acronym subsitition where if an
acronym subtitution contained an acronym (nested acro), the
nested acro would not, it self, be replaced. Have fixed this
by placing replacement code in a while loop that check for any
and all instances of acro commands in line. Thus, as long as
there are things to replace it will keep trying to replace them.

Also updates the way acronyms are replaced by creating a dictionary
where keys are regex patterns for finding acro calls and values
are subdictionaries for how to format replacements or, say long and
short forms, plural, etc. There are also sub-sub dictionaries for
certain commands defining how unused and already used acronyms
are replaced.

v0.9.2